move schema to internal
