+++ internal/object/config.go
@@ -2,7 +2,7 @@ package object
 
 import (
 	shell "github.com/ipfs/go-ipfs-api"
-	"github.com/sonr-io/sonr/pkg/schemas"
+	"github.com/sonr-io/sonr/internal/schemas"
 )
 
 type Config struct {
