+++ internal/schemas/interface.go
