+++ internal/schemas/links.go
