+++ internal/schemas/utils.go
